pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "searchstack"
version = "0.1.0"
description = "Open-source SEO / AEO / GEO tech stack for AI founders"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Hyperfocus" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "google-auth",
    "google-auth-oauthlib",
    "google-api-python-client",
    "requests",
    "tomli; python_version < '3.11'",
]

[project.scripts]
searchstack = "searchstack.cli:main"

[project.urls]
Homepage = "https://github.com/hyperfocus/searchstack"
Repository = "https://github.com/hyperfocus/searchstack"
Documentation = "https://github.com/hyperfocus/searchstack#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/searchstack"]
